+++ packages/mint-components/src/components/sqm-referral-code/useReferralCode.tsx
@@ -14,23 +14,13 @@ import {
   REFERRAL_CODES_NAMESPACE,
   SET_CODE_COPIED,
 } from "../sqm-referral-codes/useReferralCodes";
-
-interface ShareLinkProps {
-  programId?: string;
-  tooltiptext: string;
-  tooltiplifespan: number;
-  linkOverride?: string;
-}
+import { ReferralCode } from "./sqm-referral-code";
 
 const MessageLinkQuery = gql`
-  query ($programId: ID, $engagementMedium: UserEngagementMedium!) {
+  query getReferralCode($programId: ID) {
     user: viewer {
       ... on User {
-        shareLink(
-          programId: $programId
-          engagementMedium: $engagementMedium
-          shareMedium: DIRECT
-        )
+        referralCode(programId: $programId)
       }
     }
   }
@@ -42,7 +32,7 @@ const WIDGET_ENGAGEMENT_EVENT = gql`
   }
 `;
 
-export function useShareLink(props: ShareLinkProps): CopyTextViewProps {
+export function useReferralCode(props: ReferralCode): CopyTextViewProps {
   const { programId = useProgramId() } = props;
   const user = useUserIdentity();
   const engagementMedium = useEngagementMedium();
@@ -53,15 +43,14 @@ export function useShareLink(props: ShareLinkProps): CopyTextViewProps {
 
   const { data } = useQuery(
     MessageLinkQuery,
-    { programId, engagementMedium },
-    !user?.jwt || !!props.linkOverride || contextData?.shareLink !== undefined
+    { programId },
+    !user?.jwt || contextData?.referralCode !== undefined
   );
   const [sendLoadEvent] = useMutation(WIDGET_ENGAGEMENT_EVENT);
-
-  const [setCopied, copiedRes] = useMutation(SET_CODE_COPIED);
+  const [setCopied] = useMutation(SET_CODE_COPIED);
 
   const copyString =
-    (contextData?.shareLink || data?.user?.shareLink) ??
+    (contextData?.referralCode || data?.user?.referralCode) ??
     // Shown during loading
     "...";
 
@@ -72,7 +61,6 @@ export function useShareLink(props: ShareLinkProps): CopyTextViewProps {
       await setCopied({ referralCode: contextData.referralCode });
       contextData.refresh();
     }
-
     // Should well supported: https://developer.mozilla.org/en-US/docs/Web/API/Clipboard#browser_compatibility
     // Only if called from a user-initiated event
     navigator.clipboard.writeText(copyString);
@@ -92,5 +80,13 @@ export function useShareLink(props: ShareLinkProps): CopyTextViewProps {
     });
   }
 
-  return { ...props, onClick, open, copyString: copyString };
+  return {
+    ...props,
+    onClick,
+    open,
+    copyString: copyString,
+    isCopied: contextData?.isCopied,
+    showNotificationText:
+      props.showNotificationText && !!contextData?.singleUse,
+  };
 }